@ refactor(mark-verses): drop "IsMobile" from names now used on desktop too
Mark Verses is now used on desktop as well as mobile, so rename the
component files and identifiers to drop the misleading "IsMobile"
suffix:

- PassageDetailMarkVersesIsMobile.{tsx,test.tsx,cy.tsx} -> PassageDetailMarkVerses.*
- MarkVersesTableIsMobile.tsx -> MarkVersesTable.tsx
- exported components / props types renamed to match
- id "mark-verses-mobile" -> "mark-verses", table aria-label
  "mobile mark verses table" -> "mark verses table"
- update the two importers and the XLIFF sourcefile context

@
